Add files via upload
Business Case: To predict the risky driver who is more likely to claim the insurance.

Dataset description: 'target' is the dependent feature and the rest 58 are the independent features. In this dataset, features belonging to similar group are tagged with similar feature names. such as (ind,reg,car,calc)

Here, ind: is related to individual or driver, reg: is related to region car: is related to car calc: is an calculated feature.

Also, the features contain a postfix such as 'bin' and 'cat' which represents binary and categorical features respectively and the rest of the feature are either continuous or ordinal.